build(Justfile): introduce justfile
As a compile language, working with Go more or less necessiates a
"Makefile" of some sort. However, Make can arcane and hard to work with.
A more recent alternative taht appears to provide a more sensible user
interface is "Just", which will be trialed in this project.

See https://github.com/casey/just